Disable "Create Subject" button if data is incomplete
- Closely connected to the previous commit, the button for creating a new SUBJECT is getting disabled if the entered data is not complete when creating a new subject.
    - Either the name is empty
    - Or the weighting is empty or invalid (not a number)

Warning: This part doesn't fully work yet. Specifically erasing the last remaining character of the weighting (`TextBox` for the weighting) after there has been content inside, so the content is completely empty, doesn't disable the button, even though the weighting value is invalid, because it's an empty string. This is happening due to the fact that Avalonia doesn't call the `set`-accessor of the property bound to the `Text` property of the `TextBox` when the last remaining character gets erased. This should be fixed soon, as soon as I find a way to fix it. Whether that's a bug in the Avalonia framework, an intended feature or my mistake can't be said for sure yet. I suspect a bug or intended feature (hopefully with workaround) in the Avalonia framework.