Change date validation rules for day
In case the month or the year are not valid during the date validation, the day can't be exactly validated, because it's validity depends on the month and year. I covered all the cases where the day can be clearly marked as invalid, no matter if the month/year are valid or not.

- If the month is invalid:
    - The day has to be between 1 and 31 (both inclusive), because these values could be valid in some months.
- If the year is invalid:
    - If the month is equal to 2 (February), the day has to be between 1 and 29 (both inclusive), because these values could be valid in some years.
    - If the month is NOT equal to 2 (something else than February), the day's validity can be determined through `DateTime.DaysInMonth` with the valid month and the year is set to 2021, because the year doesn't matter here.