Change display when no elements are contained (Grade-list)
If the selected school year or subject doesn't contain any elements (subjects/grades), a text is displayed, stating exactly this and referring to the button, with which the user can add an element.
This display has been changed in this commit (ONLY FOR THE GRADE-LIST).

- Added an illustration of this button in a smaller format between the text.
Example: "Use the green button [ILLUSTRATION] in the bottom-right corner."
The illustration is in-between the text-block.