Fix "no-elements" text showing when it's not supposed to
Finally fixed the issue where the "no-elements"-text, displayed if there are no elements in the list, currently being viewed, was sometimes displayed when it wasn't supposed to. Now it's only being shown if the currently opened list really doesn't contain any elements and stays hidden if the list contains elements.