Fix: Closing Add-Window cleans data now
In the previous commit, I have added the functionality to close the Add-Window once the user presses the button for creating/editing an element (grade, subject or year). This has led to unwanted behaviour: After creating/editing an element and reopening the Add-Window to add another element or edit one, the data of the previously created/edited element still persisted and was applied to the text- and check-boxes of the window, instead of being empty or having the default value it usually has.
This has been fixed!

- Renamed `StopEditing` to `EraseData`, so this function doesn't indicate that it's only intended for the editing part. It's for erasing the data in the respective ViewModel in every case.
- The logic for the bugfix has been added to this virtual method (`EraseData`) and the overriden members in the deriving classes.