Bugfix: Incorrect average calculation for SchoolYear
After the last commit, the average of a year (and in the future, when partial grades will have been implemented, of a subject too) should only consider subjects(/grades) that count and have 1 or more grades inside. The latter part didn't work properly, because the produced enumerable had been tried to be converted to an array of type `IGradable` and the fallback value if this operation didn't succeed was the entire `gradables` collection without any restrictions.
This has been fixed!

I really don't know where my head was when I wrote this.