Bugfix: `TargetGradeWindow` correct button color behaviour
There are two buttons in the `TargetGradeWindow` for switching the calculation mode. The currently selected button's background is green (typical green) and the other one has a transparent background. If the user presses the button, which is NOT selected, the two buttons should swap their backgrounds.
That didn't happen though. This has been fixed now!

- Added `Color` extension methods. You can change one or multiple specific axes of a `Color` object with an extension method. There are extension methods for each possible combination (R, RG, RGB, RGBA, G, GB, GA, GBA, A etc.)