Add/Edit Grade: Validate Day when month or year change
- Imagine the following situation:
You are creating a new grade or editing a grade and you set the day to 29 and the year to 2021 (not a leap year). If the month is set to any other month than February, it's fine. If you change the month to February however, the day is actually not valid anymore, because February in a non-leap year doesn't have 29 days, but 28. Before the changes in this commit, the day's TextBox still had a green border. Same with when the year changes.
This has been changed and the day's validation state changes appropriately.

Now everything should be fine with the date, no issues! :)