Bugfix: "No elements" text not appearing sometimes
An issue has been discovered, where the `CollectionChanged` of the "Items" `ObservableCollection` of a ListView was detached after the control re-appeared (switch to another ListView and go back). This caused the check whether there are any elements or not, not to take place, which in turn made the text, that's supposed to point to this not appear and the other way around, after this happened, if you wanted to create a new element, this text wouldn't disappear until you re-enable the control. This has been fixed with...

- Removed the list, which stores opened pages, so I can reuse them. Instead, each time a page is supposed to be opened, a new instance gets created and the functionality around the list has been removed completely.