Redesign Topbar (different functionality)
Earlier, there was a clickable text in the topbar that was holding the name of the currently selected year/semester and a clickable text that was holding the name of the currently selected subject.
When the year-text had been clicked, the view returned to all years (YearList).
When the subject-text had been clicked, the view returned to the subjects of the currently selected year (SubjectList).

Logically, this would not have made much sense. If the user clicks on the "Semester 2021/22" text, you may expect the view to change to the subjects of the "Semester 2021/22" year/semester.
If the user clicks on the "English" text, you may expect the view to change to the grades of the "English" subject.

The latter logic has been implemented and made fully functional in this commit. This idea may be discarded in the near future. This is just a draft how it may look.

Changes:
- Added "..." and "/" text for changing to the list of all years. When the grey "..." gets clicked, the logic will be executed. "/" is just a separator, as always.
- Changed the Grid columns of the Grid that holds those texts.
- Changed structure, indentation and order of the properties of each TextBlock in the topbar.
- Changed the `PointerPressed` event-methods to fit the new logic and separated most of the logic into a separate generic method to allow more re-usability, make changes easier applicable to all event-methods and avoid duplicate code.
- Added a few additional null-checks and checks for the length of arrays before applying them, to errors.