Reduce stroke thickness for symbol of "No Elements" text
The text that appears when there are no elements inside the view got a little update in one of the last commits. Due to the redesign of the add-button and because this text contains a symbol representing this add-button, the symbol had been redesigned to match the current look of the add-button.
But immediately after commiting, an observation has been made, that the stroke of the ellipse is way too thick, because the `StrokeThickness` had been taken over from the original add-button, which is way bigger, so a thickness of 5 looks fine.

- The same proportion of `StrokeThickness` (5) and Width (150) of the add-button has been applied to the thickness of the symbol's ellipse (for all ListViews):
    - Width = 40
    - `StrokeThickness` = 1.333