Secure restriction for name-length if changed by hand
If the name of an element gets changed in the JSON file, which is quite hidden, but just in case, the max length shouldn't be exceeded. That's why each element-class (`Grade`, `Subject`, `SchoolYear`) has such a check in the constructor and takes a substring of [MaxNameLength] characters if the deserialized name is too long.